requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pandas
pytest
coverage
flake8
black
yapf
python-gitlab
twine
six==1.14.0
numpy==1.18.5
pandas
scikit-learn==0.22
xgboost
joblib
memoized-property
mlflow
s3fs
gcsfs
google-cloud-storage
termcolor
uvicorn
fastapi
